+++ test/test_optimization.py
@@ -459,12 +459,17 @@ def test_set_files(tmp_path, model, solver, io_api):
 def test_set_files_and_keep_files(tmp_path, model, solver, io_api):
     status, condition = model.solve(
         solver,
+        io_api=io_api,
         problem_fn=tmp_path / "problem.lp",
         solution_fn=tmp_path / "solution.sol",
         log_fn=tmp_path / "logging.log",
         keep_files=True,
     )
     assert status == "ok"
+    if io_api != "direct":
+        assert (tmp_path / "problem.lp").exists()
+        assert (tmp_path / "solution.sol").exists()
+    assert (tmp_path / "logging.log").exists()
 
 
 @pytest.mark.parametrize("solver,io_api", params)